requirements.txt
fastapi~=0.92.0
python-dotenv~=1.0.0
azure-cosmos~=4.3.1
pydantic~=1.10.5
